pyproject.toml
[project]
name = "notebooklm-clone"
version = "0.1.0"
description = "Document-grounded AI assistant with citations, memory, and podcast generation"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "fastapi==0.115.8",
  "uvicorn[standard]==0.34.0",
  "pydantic==2.10.6",
  "pydantic-settings==2.7.1",
  "python-dotenv==1.0.1",
  "python-multipart==0.0.20",
  "sqlalchemy==2.0.37",
  "alembic==1.14.1",
  "psycopg[binary]==3.2.4",
  "httpx==0.28.1",
  "passlib[bcrypt]==1.7.4",
  "python-jose[cryptography]==3.3.0",
  "email-validator==2.2.0",
  "authlib==1.4.1",
  "redis==5.2.1",
  "rq==1.16.2",
  "pymilvus==2.5.4",
  "setuptools==75.8.0",
  "numpy==2.2.2",
  "pymupdf==1.25.2",
  "trafilatura==2.0.0",
  "beautifulsoup4==4.12.3",
  "yt-dlp==2025.2.19",
  "faster-whisper==1.1.1",
  "streamlit==1.42.0",
  "requests==2.32.3",
  "pydub==0.25.1",
  "soundfile==0.13.1",
  "kokoro==0.9.4"
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.4",
  "pytest-asyncio==0.25.3",
  "pytest-cov==6.0.0",
  "mypy==1.14.1",
  "ruff==0.9.3"
]

[build-system]
requires = ["hatchling==1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/backend", "src/frontend"]

[tool.pytest.ini_options]
pythonpath = [".", "src/backend", "src/frontend"]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
plugins = []
mypy_path = ["src/backend", "src/frontend"]
files = ["src/backend/app", "src/backend/schemas", "tests"]

[[tool.mypy.overrides]]
module = [
  "requests",
  "requests.*",
  "passlib",
  "passlib.*",
  "jose",
  "jose.*",
  "pyttsx3",
  "pydub",
  "fitz",
  "bs4",
  "pymilvus",
  "soundfile",
  "kokoro",
  "faster_whisper",
  "sentence_transformers",
  "sentence_transformers.*",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "ASYNC"]
ignore = ["B008", "E501", "I001", "UP038"]
